+++ src/components/Button/style.css
@@ -1,0 +1,17 @@
+.button-content {
+  align-self: stretch;
+  display: flex;
+  flex: 0 0 auto;
+  position: relative;
+  width: 100%;
+}
+
+.button-icon {
+  align-items: flex-start;
+  display: inline-flex;
+  flex: 0 0 auto;
+  flex-direction: column;
+  margin-left: auto;
+  justify-content:center;
+}
+
