+++ src/content/examples.html
@@ -30,18 +30,18 @@
     <div class="experiment-header">
       <h1>The Examples of Components</h1>
       <div class="header-buttons">
-        <button class="button-option" id="experiment-option-btn">···</button>
-        <button class="button-option" id="experiment-info-btn">i</button>
+        <button class="button-tertiary" id="experiment-option-btn">···</button>
+        <button class="button-tertiary" id="experiment-info-btn">i</button>
       </div>
     </div>
     <h1>Component Examples</h1>
     <h2>1. Elementary Components</h2>
     <h3>1.1 Buttons and Tables</h3>
     <div class="dashboard">
-      <button class="button-submit">Submit</button>
-      <button class="button-option">Login with Google</button>
-      <button class="button-cancel">Cancel</button>
-      <button class="button-submit" disabled>Disabled Button</button>
+      <button class="button-primary">Submit</button>
+      <button class="button-tertiary">Login with Google</button>
+      <button class="button-secondary">Cancel</button>
+      <button class="button-primary" disabled>Disabled Button</button>
       <input type="text" class="input-text" placeholder="Enter here..." />
       <input
         type="text"
@@ -141,15 +141,15 @@
           maxlength="600"
           placeholder="Describe the mutation you want to apply in less than 600 chars."
         ></textarea>
-        <button class="button-submit">Submit</button>
+        <button class="button-primary">Submit</button>
       </div>
-      <div class="progress-status-bar">
+      <!-- <div class="progress-status-bar">
         <div class="filename">Filename.nc</div>
         <span>√</span>
       </div>
       <div class="accordion-item">
         Please generate N random mutants from mutating one of the residue X, Y, Z to larger residues
-      </div>
+      </div> -->
     </div>
   </body>
 </html>
