+++ src/screens/Experiments/components/ModalPanelBond/ModalPanelBond.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { TextInputDefault } from "../../../Login/components/TextInputDefault";
+import { TextInputDefault } from "../../../../components/TextInputDefault";
 import "./style.css";
 
 export const ModalPanelBond = () => {
